backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["xcfrg <30675315+xcfrg@users.noreply.github.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Flask-Cors = "^3.0.10"
Flask = "^2.1.2"
Flask-PyMongo = "^2.3.0"
python-dotenv = "^0.20.0"
Flask-JWT-Extended = "^4.4.0"
argon2-cffi = "^21.3.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
